### 2021-12-08 #3984 @whedon
[PRE REVIEW]: DCISolver.jl: A Julia Solver for Nonlinear Optimization using Dynamic Control of Infeasibility
**Submitting author:** @tmigot (<a href="http://orcid.org/0000-0001-7729-2513">Tangi Migot</a>)
**Repository:** <a href="https://github.com/JuliaSmoothOptimizers/DCISolver.jl" target ="_blank">https://github.com/JuliaSmoothOptimizers/DCISolver.jl</a>
**Version:** v0.2.4
