+++ paper/paper.bib
@@ -388,7 +388,8 @@ publisher={Multidisciplinary Digital Publishing Institute}}
 title= {Controle din{\^a}mico de infactibilidade para programa{\c{c}}{\~a}o n{\~a}o linear},
 author= {Siqueira, Abel Soares},
 year= {2013},
-school= {Universidade Estadual de Campinas}
+school= {Universidade Estadual de Campinas},
+doi = {10.47749/T/UNICAMP.2013.918998}
 }
 
 @misc{dcicpp,
