+++ paper/paper.bib
@@ -418,5 +418,6 @@ doi = {10.47749/T/UNICAMP.2013.918998}
   number={2},
   pages={201--213},
   year={2002},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s101070100263}
 }
