+++ paper/paper.bib
@@ -123,7 +123,7 @@
 @Manual{HSL,
   Title                    = {The {HSL} Mathematical Software Library},
   Author                   = {HSL},
-  Note                     = {\url{http://www.hsl.rl.ac.uk}},
+  url                      = {http://www.hsl.rl.ac.uk},
   Organization             = {STFC Rutherford Appleton Laboratory},
   Year                     = {2007}
 }
