### 2021-12-09 #3984 @tmigot
> Dear @tmigot
>
> Do you have any suggestions for potential reviewers? You can also use the list [here](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822). Please mention the GitHub users without using the @ character (just because we don't want to bother them with unnecessary notifications).
>
> Thank you in advance.

Dear @jbytecode ,

Checking interests in Julia programming and optimization, the following potential reviewers might be interested : matbesancon, aivodji, odow, degleris1, jbcaillau, justinsgray, mewilhel
