### 2021-12-09 #3984 @jbytecode
👋👋👋 Dear @matbesancon  and  @aivodji  👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**DCISolver.jl: A Julia Solver for Nonlinear Optimization using Dynamic Control of Infeasibility**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/3984).
