### 2021-12-14 #3991 @jbytecode
Dear @odow and @jbcaillau

This is the review issue. There are 20 tasks for each reviewer. Whenever an item is solved, you can check them.

You can interact with the other reviewers, the author(s), and me. You don't have to complete your review in one step, so while the authors improve the software and paper, you can continue your review with the other tasks.

Reviewers can also contribute to repository, open new issues in the target repo as well. Please mention this issue in your review in the target repo so we can keep our eyes on what is going on broadside.

Please do not hesitate to ask me anything, any time.
