### 2021-12-14 JuliaSmoothOptimizers/DCISolver.jl#70 @odow
Paper suggestion: fix margins in plots
You might want to rebuild these plots with slightly larger margins.

The top of the P in "Proportion" is cut off, so too is the "g" in "log":

Otherwise, nice job. I like the log output from the solver.
