### 2021-12-14 #3991 @odow
This is a very nice submission that was trivial to review. The paper and software are well written and a great fit for JOSS.

The only suggestion I had was to fix the margins in the plot (https://github.com/JuliaSmoothOptimizers/DCISolver.jl/issues/70), but other than that I'm happy for this to be accepted.
