### 2021-12-15 JuliaSmoothOptimizers/DCISolver.jl#70 @tmigot
Thanks @odow , this is a nice catch. I will open a PR to fix it.
