### 2021-12-15 JuliaSmoothOptimizers/DCISolver.jl#70 @odow
Related comment: is the x-axis meant to be a log scale? It looks linear...
