### 2021-12-17 JuliaSmoothOptimizers/DCISolver.jl#70 @tmigot
It is indeed a logarithmic (base 2) scale. For instance, the figure shows that there is one problem for which DCISolver required 2^15=32768 more time than Ipopt, and we can verify it:
```
using Pkg; Pkg.activate(".")
using JLD2, GR, Plots, SolverBenchmark

@load "ipopt_dcildl_82.jld2" stats # the file `ipopt_dcildl_82.jld2` in the paper folder contains a dataframes with the results of the benchmark.
solved(df) = (df.status .== :first_order)

tim_dci = stats[:dcildl][solved(stats[:dcildl]), :elapsed_time]
tim_ipopt = stats[:ipopt][solved(stats[:dcildl]), :elapsed_time]

maximum(tim_dci./tim_ipopt) # returns 32157.21805890401 which is close to 2^15.
```
