### 2021-12-17 JuliaSmoothOptimizers/DCISolver.jl#70 @odow
Ah okay, that part needs clarifying in the text then. I through you might have done log10(actual/best), but then that was far too long!

Or could you change the plotting to use a log2 axis? It'd be nicer if the x-ticks showed `2^15` etc.
