### 2021-12-18 515307d
CompatHelper: bump compat for NLPModels to 0.18, (keep existing compat)
