+++ Project.toml
@@ -18,7 +18,7 @@ SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 HSL = "0.1.0, 0.2.0"
 Krylov = "~0.7.6"
 LDLFactorizations = "0.8.1"
-NLPModels = "0.15, 0.16, 0.17"
+NLPModels = "0.15, 0.16, 0.17, 0.18"
 SolverCore = "0.2"
 SolverTools = "0.6, 0.7"
 julia = "^1.6.0"
