### 2021-12-18 #3991 @jbytecode
@tmigot - is it possible to enhance the documentation with parameters, their types, their description, return value, and examples as shown in [here](https://julia-doc.readthedocs.io/en/latest/manual/documentation/)?
