### 2021-12-20 145218f
bump solver tools 0.8.0
