+++ Project.toml
@@ -20,7 +20,7 @@ Krylov = "~0.7.6"
 LDLFactorizations = "0.8.1"
 NLPModels = "0.15, 0.16, 0.17, 0.18"
 SolverCore = "0.2"
-SolverTools = "0.6, 0.7"
+SolverTools = "0.6, 0.7, 0.8"
 julia = "^1.6.0"
 
 [extras]
