+++ Project.toml
@@ -1,7 +1,7 @@
 name = "DCISolver"
 uuid = "bee2e536-65f6-11e9-3844-e5bb4c9c55c9"
 authors = ["Abel Soares Siqueira <abel.s.siqueira@gmail.com>"]
-version = "0.2.4"
+version = "0.2.5"
 
 [deps]
 HSL = "34c5aeac-e683-54a6-a0e9-6e0fdc586c50"
