### 2021-12-22 ddae60e
improve function's doc
