+++ src/param_struct.jl
@@ -84,6 +84,50 @@ end
 
 const TR_solvers = Dict(:TR_lsmr => TR_lsmr_struct, :TR_dogleg => TR_dogleg_struct)
 
+"""
+    MetaDCI(x, y; kwargs...)
+
+Structure containing all the parameters used in the [`dci`](@ref) call.
+`x` is an intial guess, and `y` is an initial guess for the Lagrange multiplier.
+
+# Arguments
+The keyword arguments may include:
+- `atol::T=T(1e-5)`: absolute tolerance.
+- `rtol::T=T(1e-5)`: relative tolerance.
+- `ctol::T=T(1e-5)`: feasibility tolerance.
+- `unbounded_threshold::T=T(-1e5)`: below this threshold the problem is unbounded.
+- `max_eval::Integer=50000`: maximum number of cons + obj evaluations.
+- `max_time::Float64=120.0`: maximum number of seconds.
+- `max_iter::Integer=500`: maximum number of iterations.
+- `max_iter_normal_step::Integer=typemax(Int)`: maximum number of iterations in normal step.
+- `λ_struct::comp_λ_cgls=comp_λ_cgls(length(x0), length(y0), S)`.
+- `linear_solver::Symbol=:ldlfact`: Solver for the factorization. options: `:ma57`.
+- `decrease_γ::T=T(0.1)`: Regularization for the factorization: reduce `γ` if possible, `> √eps(T)`, between tangent steps.
+- `increase_γ::T=T(100.0)`: Regularization for the factorization: up `γ` if possible, `< 1/√eps(T)`, during the factorization.
+- `δmin::T=√eps(T)`: Regularization for the factorization: smallest value of `δ` used for the regularization.
+- `feas_step::Symbol=:feasibility_step`: Normal step.
+- `feas_η₁::T=T(1e-3)`: Feasibility step: decrease the trust-region radius when `Ared/Pred < η₁`.
+- `feas_η₂::T=T(0.66)`: Feasibility step: increase the trust-region radius when `Ared/Pred > η₂`.
+- `feas_σ₁::T=T(0.25)`: Feasibility step: decrease coefficient of the trust-region radius.
+- `feas_σ₂::T=T(2.0)`: Feasibility step: increase coefficient of the trust-region radius.
+- `feas_Δ₀::T=one(T)`: Feasibility step: initial radius.
+- `bad_steps_lim::Integer=3`: Feasibility step: consecutive bad steps before using a second order step.
+- `feas_expected_decrease::T=T(0.95)`: Feasibility step: bad steps are when `‖c(z)‖ / ‖c(x)‖ >feas_expected_decrease`.
+- `TR_compute_step::Symbol=:TR_lsmr`: Compute the direction in feasibility step: options: `:TR_dogleg`.
+- `TR_struct::Union{TR_lsmr_struct, TR_dogleg_struct}=TR_lsmr_struct(length(x0), length(y0), S)`.
+- `compρ_p1::T=T(0.75)`: update ρ as `ρ = max(min(ngp, p1) * ρmax, ϵ)`.
+- `compρ_p2::T=T(0.90)`: update ρ as `ρ = primalnorm * p2` if not sufficiently feasible.
+- `ρbar::T=T(2.0)`: radius of the larger cylinder is `ρbar * ρ`.
+- `tan_Δ::T=one(T)`: Tangent step trust-region parameters: initial trust-region radius.
+- `tan_η₁::T=T(1e-2)`: Tangent step trust-region parameters: decrease the trust-region radius when `Ared/Pred < η₁`.
+- `tan_η₂::T=T(0.75)`: Tangent step trust-region parameters: increase the trust-region radius when `Ared/Pred > η₂`.
[…]
