### 2021-12-22 #3991 @tmigot
@jbytecode @odow Thanks for the suggestion. I opened a PR that improve the documentation of the main functions.
Would that answer your question?
