### 2021-12-22 JuliaSmoothOptimizers/DCISolver.jl#70 @tmigot
@odow I changed in #73  the xticks to show the log2 basis.
![ipopt_dcildl_82](https://user-images.githubusercontent.com/25304288/147158035-0e1d93cb-cf24-4db0-b6f1-24161eddb2d4.png)
