### 2021-12-22 #3991 @jbytecode
looks better. thank you for adding documentation. but I think there is still some room to improve the documentation. you can add
- function definition
- arguments, their types, and a description
- a minimal example or more
- references

to each function. if you need help please have a look at the popular repos, if you still need help I can suggest some.

thanks.
