### 2021-12-22 #3991 @tmigot
Ok, I updated only the external functions in the [PR 74](https://github.com/JuliaSmoothOptimizers/DCISolver.jl/pull/74). I will also update the internal functions in the same way.
