### 2021-12-22 #3991 @jbytecode
here is an example:

https://github.com/jbytecode/LinRegOutliers/blob/master/src/hs93.jl
