+++ src/DCISolver.jl
@@ -21,22 +21,27 @@ include("main.jl")
     dci(nlp, x; kwargs...)
     dci(nlp, meta, x)
 
-Compute a local minimum of an equality-constrained optimization problem `nlp<:AbstractNLPModel`, see `NLPModels.jl`.
-The returned value is a `GenericExecutionStats`, see `SolverCore.jl`.
+Compute a local minimum of an equality-constrained optimization problem using DCI algorithm from Bielschowsky & Gomes (2008).
 
-If `x` is not specified, then `nlp.meta.x0` is used as an initial guess.
-The keyword arguments are used to initialize a [`MetaDCI`](@ref).
+# Arguments
+- `nlp::AbstractNLPModel`: the model solved, see `NLPModels.jl`.
+- `x`: Initial guess. If `x` is not specified, then `nlp.meta.x0` is used.
+- `meta`: The keyword arguments are used to initialize a [`MetaDCI`](@ref).
 
 For advanced usage, the principal call to the solver uses a [`DCIWorkspace`](@ref).
 
     dci(nlp, meta, workspace)
 
+# Output
+The returned value is a `GenericExecutionStats`, see `SolverCore.jl`.
+
+# References
 This method implements the Dynamic Control of Infeasibility for
 equality-constrained problems described in
 
     Dynamic Control of Infeasibility in Equality Constrained Optimization
     Roberto H. Bielschowsky and Francisco A. M. Gomes
-    SIAM J. Optim., 19(3), 1299–1325.
+    SIAM J. Optim., 19(3), 2008, 1299–1325.
     https://doi.org/10.1137/070679557
 
 # Examples
@@ -44,6 +49,7 @@ equality-constrained problems described in
 julia> using DCISolver, ADNLPModels
 julia> nlp = ADNLPModel(x -> 100 * (x[2] - x[1]^2)^2 + (x[1] - 1)^2, [-1.2; 1.0]);
 julia> stats = dci(nlp)
+"Execution stats: first-order stationary"
 ```
 """
 function dci(nlp::AbstractNLPModel, x::AbstractVector{T}; kwargs...) where {T}
@@ -60,11 +66,10 @@ function dci(
 end
 dci(nlp::AbstractNLPModel; kwargs...) = dci(nlp, nlp.meta.x0; kwargs...)
 
-"""compute_gBg
-  B is a symmetric sparse matrix
-  whose lower triangular given in COO: (rows, cols, vals)
-
-  Compute ∇ℓxλ' * B * ∇ℓxλ
+"""
+    compute_gBg(nlp, rows, cols, vals, ∇ℓzλ)
+  
+Compute `gBg = ∇ℓxλ' * B * ∇ℓxλ`, where `B` is a symmetric sparse matrix whose lower triangular is given in COO-format.
 """
 function compute_gBg(
   nlp::AbstractNLPModel,
@@ -74,7 +79,7 @@ function compute_gBg(
   ∇ℓzλ::AbstractVector{T},
 ) where {T}
   gBg = zero(T)
-  for k = 1:(nlp.meta.nnzh)
+  for k = 1:(nlp.meta.nnzh) # TODO: use |vals| and remove dependency in nlp
     i, j, v = rows[k], cols[k], vals[k]
     #v * ∇ℓxλ[i] * ∇ℓxλ[j] * (i == j ? 1 : 2)
     gBg += v * ∇ℓzλ[i] * ∇ℓzλ[j] * (i == j ? 1 : 2)
@@ -83,10 +88,9 @@ function compute_gBg(
 end
 
 """
-`regularized_coo_saddle_system!(nlp, rows, cols, vals, γ = γ, δ = δ)`
[…]
