+++ src/dci_tangent.jl
@@ -1,18 +1,32 @@
-"""
-min q(d):=¹/₂dᵀBd + dᵀg
-s.t Ad = 0
-    ‖d‖ ≦ Δ
-where B is an approximation of hessian of the Lagrangian, A is the jacobian
-matrix and `g` is the projected gradient.
+@doc raw"""
+    (z, cz, fz, status, Δ, Δℓ, γ, δ) = tangent_step!(nlp, z, λ, cz, normcz, fz, LDL, vals, g, ℓzλ, gBg, ρ, γ, δ, meta, workspace; kwargs...)
 
-Return status with outcomes:
-- :cauchy_step, :newton, :dogleg,
-- :unknown if we didn't enter the loop.
-- :small_horizontal_step
-- :tired if we stop due to max_eval or max_time
-- :success if we computed z such that ‖c(z)‖ ≤ meta.ρbar * ρ and Δℓ ≥ η₁ q(d)
+Solve the following problem
+```math
+\begin{aligned}
+\min_{d} \quad & q(d):=\frac{1}{2} d^T B d + d^T g \\
+\text{s.t.} \quad & Ad = 0,
+& \|d\| \leq \Delta,
+\end{aligned}
+```
+where `B` is an approximation of the Hessian of the Lagrangian, `A` is the jacobian
+matrix, and `g` is the projected gradient.
 
-See https://github.com/JuliaSmoothOptimizers/SolverTools.jl/blob/78f6793f161c3aac2234aee8a27aa07f1df3e8ee/src/trust-region/trust-region.jl#L37
+# Output
+- `z`, `cz`, and `fz`: the new iterate, and updated evaluations.
+- `status`: computation status.
+- `Δ`: trust-region parameter.
+- `Δℓ`: differential in Lagrangian computation.
+- `γ`, `δ`: updated values for the regularization parameters.
+
+Return `status` with possible outcomes:
+- `:cauchy_step`, `:newton`, `:dogleg` if successful step.
+- `:unknown` if we didn't enter the loop.
+- `:small_horizontal_step`.
+- `:tired` if we stop due to `max_eval` or `max_time`.
+- `:success` if we computed `z` such that `‖c(z)‖ ≤ meta.ρbar * ρ` and `Δℓ ≥ η₁ q(d)`.
+
+See [`SolverTools.jl`](https://github.com/JuliaSmoothOptimizers/SolverTools.jl/blob/78f6793f161c3aac2234aee8a27aa07f1df3e8ee/src/trust-region/trust-region.jl#L37)
 for `SolverTools.aredpred`
 """
 function tangent_step!(
@@ -123,13 +137,30 @@ function tangent_step!(
   return z, cz, fz, status, Δ, Δℓ, γ, δ #ℓzλ
 end
 
-"""
-Compute a direction `d` with three possible outcomes:
+@doc raw"""
+    (d, dBd, status, γ, δ, vals) = compute_descent_direction!(nlp::AbstractNLPModel, gBg, g, Δ, LDL, γ, δ, vals, d, meta, workspace)
+
+Compute a direction `d` solution of
+```math
+\begin{aligned}
+\min_{d} \quad & q(d) \\
+\text{s.t.} \quad & \|d\| \leq \Delta.
+\end{aligned}
+```
+
+# Output
+The returned arguments are:
+- `d`: the computed direction.
+- `dBd`: updated scalar product.
+- `status`: computation status.
+- `γ`, `δ`: updated values for the regularization parameters.
+- `vals`: update values for the SQD system.
+
[…]
