+++ src/workspace.jl
@@ -2,6 +2,7 @@
     DCIWorkspace(nlp, meta, x)
 
 Pre-allocate the memory used during the [`dci`](@ref) call.
+Returns a `DCIWorkspace` structure.
 """
 struct DCIWorkspace{T, S <: AbstractVector{T}, Si <: AbstractVector{<:Integer}, Op}
   x0::S
