### 2021-12-23 #3991 @jbytecode
one more thing:

Why are there some functions in test case, e.g., [here](https://github.com/JuliaSmoothOptimizers/DCISolver.jl/blob/e1dddf1c0b5bc8be2b4070b18290e4070a194aa5/test/runtests.jl#L101)

we can use nested test cases like

```julia
@testset "...." begin
   @testset "...." begin
       ....
   end
end
```

so you don't need to wrap more than one `testcase` in a function and then call this function manually in test environments. This is unusual. Please tell me if there is a specific reason to do that.

Thanks.
