### 2021-12-23 #3991 @tmigot
Thanks for the suggestion, there was indeed no specific reason for that. I opened a PR to fix this.
