### 2022-01-13 #3991 @jbcaillau
@tmigot @jbytecode Hi, I’ve read the paper and tried the code:
- very clear paper, well motivated
- DCI is definitely a nice addition
- the comparison with state of the art (though more general math prog oriented code / ipopt) on a standard benchmark is convincing & promising
- installation & example are OK
- still need to run the benchmark myself for a check
- I’ve seen that @jbytecode pointed some missing info in the code doc; nothing else from my side
- minor request: the provided example is very simple, which is OK; a richer example in notebook format would be a nice addition
