### 2022-01-14 #3991 @jbytecode
it seems a community guidelines part is missing, right?

@tmigot please consider the suggestions and apply the needes stuff if you are agreed with them. Please ping us when you have done. Thank you in advance
