### 2022-01-14 #3991 @jbytecode
Dear @tmigot

Since there is an item in the reviewers' item list that states

`Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support`

please add some related sections in your README.md of repo.

Thank you in advance.
