+++ docs/Project.toml
@@ -1,16 +1,24 @@
 [deps]
 ADNLPModels = "54578032-b7ea-4c30-94aa-7cbd1cce6c9a"
 CUTEst = "1b53aba6-35b6-5f92-a507-53c67d53f819"
+CaNNOLeS = "5a1c9e79-9c58-5ec0-afc4-3298fdea2875"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 GR = "28b8d3ca-fb5f-59d9-8090-bfdbd6d07a71"
+Gridap = "56d4f2e9-7ea1-5844-9cf6-b9c51ca7ce8e"
 JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
+JSOSolvers = "10dff2fc-5484-5881-a0e0-c90441020f8a"
+LDLFactorizations = "40e66cde-538c-5869-a4ad-c39174c6795b"
+Literate = "98b081ad-f1c9-55d3-8b20-4c87d4299306"
+Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
 NLPModels = "a4795742-8479-5a88-8948-cc11e1c8c1a6"
 NLPModelsIpopt = "f4238b75-b362-5c4c-b852-0801c9a21d71"
+NLPModelsModifiers = "e01155f1-5c6f-4375-a9d8-616dd036575f"
+PDENLPModels = "80da258d-c172-4540-9646-64f172512520"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 SolverBenchmark = "581a75fa-a23a-52d0-a590-d6201de2218a"
 SolverCore = "ff4d7338-4cf1-434d-91df-b86cb86fb843"
 
 [compat]
-ADNLPModels = "0.2"
+ADNLPModels = "0.3"
 Documenter = "0.26"
