+++ previews/PR78/example.jl
@@ -1,6 +1,6 @@
 using Gridap, PDENLPModels
 
-n = 100
+n = 20
 domain = (-1, 1, -1, 1)
 partition = (n, n)
 model = CartesianDiscreteModel(domain, partition)
