+++ docs/Project.toml
@@ -20,5 +20,15 @@ SolverBenchmark = "581a75fa-a23a-52d0-a590-d6201de2218a"
 SolverCore = "ff4d7338-4cf1-434d-91df-b86cb86fb843"
 
 [compat]
-ADNLPModels = "0.3"
+ADNLPModels = "0.3.1"
+CUTEst = "0.12.2"
 Documenter = "0.26"
+Gridap = "0.15.5"
+JSOSolvers = "0.7.4"
+LDLFactorizations = "0.8.1"
+NLPModels = "0.18.1"
+NLPModelsIpopt = "0.8.0"
+NLPModelsModifiers = "0.5.1"
+PDENLPModels = "0.3.0"
+SolverBenchmark = "0.5.0"
+SolverCore = "0.2.2"
