### 2022-01-18 #3991 @tmigot
Hi @jbcaillau @jbytecode
I tried to add some new features following your recommendations:
- PR 77 added contribution guidelines in the readme
- PR 78 added a tutorial with a discretized PDE-constrained optimization problem. I cannot really push the size of the problem, because everytime there is a commit in the package, the scripts in the docs are re-build and I believe it would be counter-productive to spend 1hrs+ everytime there. However, the size of the problem follows a discretization parameter clearly identified.
- PR 80 added a badge/url to binder hosting the notebook corresponding to this tutorial.

The link to the new tutorial: https://juliasmoothoptimizers.github.io/DCISolver.jl/dev/example/

@jbcaillau to re-run the benchmark from the paper, you can:
- clone the repository
- move to the branch `joss-paper-branch` ; then move to the `paper` folder.
- run `benchmark.jl` which will generate a .jld2 file with the result, and finally run `figures.jl` to generate the figures.
