+++ dev/example.jl
@@ -58,14 +58,15 @@ reset!(nlp);
 using DCISolver, Logging
 
 stats_dci = with_logger(NullLogger()) do
-    dci(nlp, stats_trunk.solution, atol = 1e-5, rtol = 0.0)
+  dci(nlp, stats_trunk.solution, atol = 1e-5, rtol = 0.0)
 end
 
 stats_dci.counters
 
 stats_ipopt.elapsed_time, stats_dci.elapsed_time
 
-(stats_ipopt.objective, stats_ipopt.primal_feas, stats_ipopt.dual_feas), (stats_dci.objective, stats_dci.primal_feas, stats_dci.dual_feas)
+(stats_ipopt.objective, stats_ipopt.primal_feas, stats_ipopt.dual_feas),
+(stats_dci.objective, stats_dci.primal_feas, stats_dci.dual_feas)
 
 # This file was generated using Literate.jl, https://github.com/fredrikekre/Literate.jl
 
