### 2022-01-18 #3991 @jbcaillau
@tmigot thanks for the update!
- very neat notebook example
- on `binder`, I don’t see any `toml` file indicating the dependencies: how is this suppose to work?
- BTW, clicking the `binder` link launches the notebook but no kernel is found
