+++ docs/src/Project.toml
@@ -1,0 +1,18 @@
+[deps]
+Gridap = "56d4f2e9-7ea1-5844-9cf6-b9c51ca7ce8e"
+JSOSolvers = "10dff2fc-5484-5881-a0e0-c90441020f8a"
+LDLFactorizations = "40e66cde-538c-5869-a4ad-c39174c6795b"
+Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
+NLPModels = "a4795742-8479-5a88-8948-cc11e1c8c1a6"
+NLPModelsIpopt = "f4238b75-b362-5c4c-b852-0801c9a21d71"
+NLPModelsModifiers = "e01155f1-5c6f-4375-a9d8-616dd036575f"
+PDENLPModels = "80da258d-c172-4540-9646-64f172512520"
+
+[compat]
+Gridap = "0.15.5"
+JSOSolvers = "0.7.4"
+LDLFactorizations = "0.8.1"
+NLPModels = "0.18.1"
+NLPModelsIpopt = "0.8.0"
+NLPModelsModifiers = "0.5.1"
+PDENLPModels = "0.3.0"
