### 2022-01-19 #3991 @tmigot
Ok, I finally figured it out, the link should be working with Julia now. It is slightly tricky to get the environment, so I added a link toward the Project.toml and the Manifest.toml at the beginning of the carnet.
