### 2022-01-20 #3991 @jbcaillau
@tmigot well done! see also below to speed up the process (double repo)
https://discourse.jupyter.org/t/tip-speed-up-binder-launches-by-pulling-github-content-in-a-binder-link-with-nbgitpuller/922
