### 2022-01-27 #3991 @jbcaillau
@jbytecode @tmigot Hi, thank you for your patience:
- wanted to run some tests before checking items in the list, and had some trouble with Julia 1.6.x / 1.7
- I am not sure of what is expected regarding item « Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified? » so I left it unchecked

All in all, this is a very nice and welcome piece of work so I do not want to delay its publication any longer on JOSS. Well done, @tmigot *et al*!
