### 2022-01-27 #3991 @jbytecode
@jbcaillau what kind of trouble you had, could you please clarify?

Submissions on JOSS requires software testing in any means, btw, in our case, this is a Julia package, so we expect

julia>]
test DCISolver

to pass all of the tests in Julia REPL. Please check.

And, of course, the test directory should be capable of testing the whole of the functionality that the package serves.
