### 2022-01-27 #3991 @jbytecode
@tmigot In the README.md file the title "Bug reports and discussions" is okay, however, please add some information here on how to contribute to software by means of adding new functionality, improving the methods, etc. because the current information is only about reporting bugs. Thank you in advance.
