### 2022-01-27 #3991 @tmigot
> @tmigot In the README.md file the title "Bug reports and discussions" is okay, however, please add some information here on how to contribute to software by means of adding new functionality, improving the methods, etc. because the current information is only about reporting bugs. Thank you in advance.

Thank you for the suggestion. Could you be more specific?

We wrote `Focused suggestions and requests can also be opened as issues. Before opening a pull request, start an issue or a discussion on the topic, please.` in this spirit. The idea is to open an issue to discuss new functionalities.

If you mean to specify the future projects for the package, we also use the Github Projects.
