### 2022-02-02 #3991 @jbcaillau
@jbytecode Hi, test passed too on 1.7, see below
```julia
Testing Running tests...
[ Info: libhsl_ma57 not defined.
Test Summary:             | Pass  Total
Positive definite systems |    3      3
Test Summary:      | Pass  Total
Indefinite systems |    4      4
Test Summary:        | Pass  Total
Failed factorization |    2      2
Test Summary: | Pass  Total
SQD system    |    4      4
Test Summary:                         | Pass  Total
LDLFactorizations with regularization |    3      3
[ Info:           stage    iter   #f+#c      f(x)         ℓ      ‖∇L‖    ‖c(x)‖      ρmax         ρ           status       ‖d‖         Δ
[ Info:            init       0       2   0.0e+00   0.0e+00   1.4e+00   0.0e+00   7.1e+01       NaN                -         -         -
[ Info:              Tr       0       4  -1.4e+00  -1.4e+00         -   0.0e+00         -         -          success   1.0e+00   2.0e+00
[ Info:               T       0       4  -1.4e+00  -1.4e+00   1.4e+00   0.0e+00   7.1e+01   4.1e+01          success         -         -
[ Info:              Tr       0       6  -3.0e+01  -3.0e+01         -   0.0e+00         -         -          success   2.0e+01   4.0e+01
[ Info:               T       1       6  -3.0e+01  -3.0e+01   1.4e+00   0.0e+00   7.1e+01   4.1e+01          success         -         -
[ Info:              Tr       0       8  -6.0e+02  -6.0e+02         -   0.0e+00         -         -          success   4.0e+02   8.0e+02
[ Info:               T       2       8  -6.0e+02  -6.0e+02   1.4e+00   0.0e+00   7.1e+01   4.1e+01          success         -         -
[ Info:              Tr       0      10  -1.2e+04  -1.2e+04         -   0.0e+00         -         -          success   8.0e+03   1.6e+04
[ Info:               T       3      10  -1.2e+04  -1.2e+04   1.4e+00   0.0e+00   7.1e+01   4.1e+01          success         -         -
[ Info:              Tr       0      12  -2.4e+05  -2.4e+05         -   0.0e+00         -         -          success   1.6e+05   3.2e+05
[ Info:               T       4      12  -2.4e+05  -2.4e+05   1.4e+00   0.0e+00   7.1e+01   4.1e+01          success         -         -
Test Summary:   | Pass  Total
Unbounded tests |    1      1

Test Summary:       | Pass  Total
Unconstrained tests |   20     20
Test Summary:                       | Pass  Total
Small equality constrained problems |   28     28
Test Summary: | Pass  Total
HS7           |    4      4
stats.solution = [4.601595018913945, 1.9558435484179804]
Test Summary: | Pass  Total
HS8           |    3      3
Test Summary: | Pass  Total
HS9           |    3      3
stats.solution = [0.9976722156536222, 0.9976721869489701, 1.0023156818519954]
Test Summary: | Pass  Total
HS26          |    3      3
stats.solution = [-1.0, 1.0000000000000075, -3.988238127092249e-9]
Test Summary: | Pass  Total
HS27          |    3      3
[ Info:               F       0       1         -         -         -   1.0e+00         -         -        too_small   0.0e+00   1.0e+00
[ Info:          F-safe       0       2         -         -         -   1.0e+00         -         -   agressive_fail   0.0e+00   1.0e+00
[ Info:               F       0       4         -         -         -   1.0e+00         -         -         reduce_Δ   1.0e+00   2.5e-01
[ Info:               F       1       5         -         -         -   1.0e+00         -         -         reduce_Δ   2.5e-01   6.2e-02
[ Info:               F       2       6         -         -         -   1.0e+00         -         -         reduce_Δ   6.2e-02   1.6e-02
[ Info:               F       3       7         -         -         -   1.0e+00         -         -         reduce_Δ   1.6e-02   3.9e-03
[ Info:               F       4       8         -         -         -   1.0e+00         -         -          success   3.9e-03   3.9e-03
[ Info:               F       5       9         -         -         -   1.0e+00         -         -          success   3.9e-03   3.9e-03
[ Info:               F       6      10         -         -         -   1.0e+00         -         -          success   3.9e-03   3.9e-03
[ Info:          F-safe       6      11         -         -         -   1.0e+00         -         -          success   1.1e-02   3.9e-03
[ Info:               F       7      12         -         -         -   1.0e+00         -         -          success   3.9e-03   7.8e-03
[ Info:          F-safe       7      13         -         -         -   1.0e+00         -         -          success   3.9e-03   7.8e-03
[ Info:               F       8      14         -         -         -   1.0e+00         -         -          success   7.8e-03   1.6e-02
[ Info:          F-safe       8      15         -         -         -   1.0e+00         -         -          success   7.8e-03   1.6e-02
[ Info:               F       9      16         -         -         -   1.0e+00         -         -          success   1.6e-02   3.1e-02
[ Info:          F-safe       9      17         -         -         -   1.0e+00         -         -          success   1.6e-02   3.1e-02
[ Info:               F      10      18         -         -         -   1.0e+00         -         -          success   3.1e-02   6.2e-02
[ Info:          F-safe      10      19         -         -         -   1.0e+00         -         -          success   3.1e-02   6.2e-02
[ Info:               F      11      20         -         -         -   1.0e+00         -         -          success   6.2e-02   1.2e-01
[ Info:          F-safe      11      21         -         -         -   1.0e+00         -         -          success   6.3e-02   1.2e-01
[ Info:               F      12      22         -         -         -   9.8e-01         -         -          success   1.2e-01   2.5e-01
[ Info:          F-safe      12      23         -         -         -   9.8e-01         -         -          success   1.3e-01   2.5e-01
[ Info:               F      13      24         -         -         -   9.4e-01         -         -          success   2.5e-01   5.0e-01
[ Info:          F-safe      13      25         -         -         -   9.4e-01         -         -          success   2.9e-01   5.0e-01
[ Info:               F      14      26         -         -         -   7.1e-01         -         -          success   5.0e-01   1.0e+00
[ Info:               F      15      27         -         -         -   7.1e-01         -         -         reduce_Δ   6.5e-01   2.5e-01
[ Info:               F      16      28         -         -         -   3.2e-02         -         -          success   1.8e-01   2.5e-01
Test Summary: | Pass  Total
Example 1     |    2      2
[ Info:               F       0       1         -         -         -   1.0e+00         -         -        too_small   0.0e+00   1.0e+00
[ Info:          F-safe       0       2         -         -         -   1.0e+00         -         -   agressive_fail   0.0e+00   1.0e+00
[ Info:               F       0       4         -         -         -   6.0e-03         -         -          success   1.0e+00   2.0e+00
Test Summary:                   | Pass  Total
Example 2, Mission: Impossible  |    2      2
     Testing DCISolver tests passed
```
