### 2022-02-02 #3991 @jbcaillau
@tmigot Still having an issue on 1.7 to run the benchmark; problem to precompile Ipopt, see
```math
ERROR: LoadError: Failed to precompile Ipopt [b6b21f68-93f8-5de0-b562-5493be1d77c9] to /home/caillau/.julia/compiled/v1.7/Ipopt/jl_NNHi1z.
```
