### 2022-02-03 #3991 @tmigot
Hi @jbcaillau !
Can you give me more information on what you are trying to run exactly? Which versions of the packages?
This seems to be unrelated to DCISolver, but with Ipopt.
