### 2022-02-09 #3991 @jbcaillau
@tmigot @jbytecode Hi guys, sorry for being unclear on this:
- from my side the review is done and the paper / package ready for JOSS (tests are OK, see post [above](https://github.com/openjournals/joss-reviews/issues/3991#issuecomment-1028413238)
- running the *benchmark* on my usual linux server raises some issues that are, as point by @tmigot, related to my Ipopt install, not to DCI. Closed.

Regards

> Hi @jbcaillau !
> Can you give me more information on what you are trying to run exactly? Which versions of the packages?
> This seems to be unrelated to DCISolver, but with Ipopt.
