### 2022-02-10 f5466cc
add citation to Julia, format numbers
