### 2022-02-10 #3991 @jbytecode
Dear @tmigot

- [ ] I sent a pull request (https://github.com/JuliaSmoothOptimizers/DCISolver.jl/pull/86) that includes adding a citation and formatting numbers. If you agree with them, please apply.

- [ ] Please re-check the whole paper by means of both language and format.

- [ ] make a tagged release and archive (e.g. with [Zenodo](https://zenodo.org/)) as described [here](<https://joss.readthedocs.io/en/latest/submitting.html#the-review-process>), and report the version number and archive DOI in this thread. Please verify that the archive deposit has the **correct metadata** (**title and author list**), or edit these if that is not the case.

- [ ] use the format vx.y.z as your versioning format.

Thank you in advance.
