### 2022-02-10 #3991 @jbytecode
your version schema is correct.

it seems the zenodo archive is a little bit outdated as it has been created in December 2021. Does thIs repo include latest changes that you applied during the review?

if its content is outdated, please create a new archive and provide the required information.

if it is okay as it is, please inform me.
