### 2022-02-10 #3991 @tmigot
Ok, I see. I can make a new release to v0.2.6 so it will be up-to-date.
