### 2022-02-10 #3991 @tmigot
Here is the doi of the version 0.2.6: 10.5281/zenodo.6037298

https://zenodo.org/record/6037298#.YgU-X4TMKUk
