### 2022-02-10 #3991 @kthyng
Hi @tmigot! Just a few things to wrap up:

1. can you update the metadata in your zenodo archive so the title and author list exactly match your JOSS paper? This gives nice continuity across the snapshot in time the paper represents.
2. Orban et al 2021 in the paper references I think should have "julia" capitalized? You can preserve capitalization with {} around characters in your .bib file. Please check the rest of your references for this too.
