+++ paper/paper.bib
@@ -113,7 +113,7 @@
 
 @Misc{orban-hsl-2021,
   author = {D. Orban and {contributors}},
-  title = {{HSL.jl}: A Julia interface to the {HSL} Mathematical Software Library},
+  title = {{HSL.jl}: A {Julia} interface to the {HSL} Mathematical Software Library},
   month = {March},
   url = {https://github.com/JuliaSmoothOptimizers/HSL.jl},
   year = {2021},
