+++ CITATION.cff
@@ -1,0 +1,44 @@
+# This CITATION.cff file was generated with cffinit.
+# Visit https://bit.ly/cffinit to generate yours today!
+
+cff-version: 1.2.0
+title: >-
+  DCISolver.jl: A Julia Solver for Nonlinear
+  Optimization using Dynamic Control of Infeasibility
+message: >-
+  If you use this software, please cite it using the
+  metadata from this file.
+type: software
+authors:
+  - given-names: Tangi
+    family-names: Migot
+    email: tangi.migot@gmail.com
+    affiliation: >-
+      GERAD and Department of Mathematics and
+      Industrial Engineering, Polytechnique Montréal,
+      QC, Canada.
+    orcid: 'https://orcid.org/0000-0001-7729-2513'
+  - given-names: Dominique
+    family-names: Orban
+    email: dominique.orban@gerad.ca
+    orcid: 'https://orcid.org/0000-0002-8017-7687'
+    affiliation: >-
+      GERAD and Department of Mathematics and
+      Industrial Engineering, Polytechnique Montréal,
+      QC, Canada
+  - affiliation: 'Netherlands eScience Center, Amsterdam, NL'
+    orcid: 'https://orcid.org/0000-0003-4451-281X'
+    email: abel.s.siqueira@gmail.com
+    given-names: Abel Soares
+    family-names: Siqueira
+identifiers:
+  - description: Zenodo archive
+    type: doi
+    value: 10.5281/zenodo.4655145
+keywords:
+  - Nonlinear Optimization
+  - Julia
+  - Nonlinear Programming
+license: MPL-2.0
+repository-code: >-
+  https://github.com/JuliaSmoothOptimizers/DCISolver.jl
