+++ CITATION.cff
@@ -29,8 +29,8 @@ authors:
   - affiliation: 'Netherlands eScience Center, Amsterdam, NL'
     orcid: 'https://orcid.org/0000-0003-4451-281X'
     email: abel.s.siqueira@gmail.com
-    given-names: Abel Soares
-    family-names: Siqueira
+    given-names: Abel
+    family-names: Soares Siqueira
 identifiers:
   - description: Zenodo archive
     type: doi
