+++ v0.2.6-alpha/siteinfo.js
@@ -1,0 +1,1 @@
+var DOCUMENTER_CURRENT_VERSION = "v0.2.6-alpha";
