### 2022-02-10 #3991 @tmigot
We updated the Zenodo page: https://zenodo.org/record/6040222#.YgWdf4TMKUk
This is a v0.2.6-alpha release as we took the chance to add a CITATION.cff in the repo.

Thanks @kthyng  for the comments.
