### 2022-02-11 #3991 @tmigot
Yes, for some reason the version title was v0.2.6-cff (meaning the one you see on github's interface) while the version name was v0.2.6-alpha. So, I fixed the title to be v0.2.6-alpha.
