### 2022-02-11 #3991 @kthyng
@jbytecode my understanding was that @tmigot wanted to incorporate the last few changes into the zenodo archive and also needed that update to change the title/authors. @tmigot did I understand correctly? Which zenodo archive and version do you want associated with your JOSS paper?
