### 2022-02-11 #3991 @tmigot
The v0.2.6 and v0.2.6-alpha are the same. Since you suggest changing the title and author, we decided to do this automatically by adding a CITATION.cff file in the repository.
So, if you can do v0.2.6-alpha for Zenodo and v0.2.6 for the repo (as the v0.2.6-alpha is not a Julia version) that would be perfect.
