### 2022-02-11 #3991 @tmigot
The doi for the v0.2.6 can be set as 10.5281/zenodo.6040222
