### 2022-02-11 #3991 @tmigot
Thanks @kthyng @jbytecode @odow @jbcaillau it's the first time I submitted to JOSS and it was a great experience!
