### 2022-02-11 #3991 @jbytecode
@tmigot do not forget to add the new citation info in your repo, and great hopes you to have infinite number of cites to your paper! :) thank you all @odow and @jbcaillau
